scripts/requirements.txt
# Python dependencies for local website tooling (not for Vercel runtime)

# R2 Storage
boto3>=1.28.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
pillow-heif>=0.16.0
pillow-heif>=0.16.0

# Machine learning for color extraction
scikit-learn>=1.3.0

# Google Photos API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
